+++ wadlib/__init__.py
@@ -6,6 +6,7 @@ from .exceptions import (
     TruncatedWadError,
     WadFormatError,
 )
+from .lumps.decorate import DecorateActor, DecorateLump
 from .validate import InvalidLumpError
 from .wad import WadFile
 from .writer import WadWriter
@@ -13,6 +14,8 @@ from .writer import WadWriter
 __all__ = [
     "BadHeaderWadException",
     "CorruptLumpError",
+    "DecorateActor",
+    "DecorateLump",
     "InvalidDirectoryError",
     "InvalidLumpError",
     "LumpInfo",